+++ collector/src/main/java/com/bitmovin/analytics/features/Feature.kt
@@ -1,9 +1,6 @@
 package com.bitmovin.analytics.features
 
-import com.bitmovin.analytics.utils.DataSerializer
-import kotlin.reflect.KClass
-
-abstract class Feature<TConfig : FeatureConfig>(val name: String, private val configClass: KClass<TConfig>) {
+abstract class Feature<TConfigContainer, TConfig : FeatureConfig>() {
     var isEnabled = true
         private set
     var config: TConfig? = null
@@ -12,18 +9,18 @@ abstract class Feature<TConfig : FeatureConfig>(val name: String, private val co
     open fun configured(authenticated: Boolean, config: TConfig?) {}
     open fun enabled() {}
     open fun disabled() {}
+    open fun reset() {}
 
     fun disable() {
         isEnabled = false
         disabled()
     }
 
-    fun configure(authenticated: Boolean, configString: String?): TConfig? {
-        if (configString != null) {
-            try {
-                config = DataSerializer.deserialize(configString, configClass.java)
-            } catch (ignored: Throwable) {
-            }
+    abstract fun extractConfig(featureConfigs: TConfigContainer): TConfig?
+
+    fun configure(authenticated: Boolean, configContainer: TConfigContainer?): TConfig? {
+        if (configContainer != null) {
+            config = extractConfig(configContainer)
         }
         configured(authenticated, config)
         return config