+++ collector/src/main/java/com/bitmovin/analytics/features/errordetails/ErrorDetailTrackingConfig.kt
@@ -2,4 +2,4 @@ package com.bitmovin.analytics.features.errordetails
 
 import com.bitmovin.analytics.features.FeatureConfig
 
-class ErrorDetailTrackingConfig : FeatureConfig()
+data class ErrorDetailTrackingConfig(override val enabled: Boolean = false, val numberOfSegments: Int? = null) : FeatureConfig