+++ collector/src/main/java/com/bitmovin/analytics/features/errordetails/OnErrorDetailEventListener.kt
@@ -1,5 +1,5 @@
 package com.bitmovin.analytics.features.errordetails
 
 interface OnErrorDetailEventListener {
-    fun onError(timestamp: Long, code: Int?, message: String?, throwable: Throwable?)
+    fun onError(code: Int?, message: String?, throwable: Throwable?)
 }