+++ collector/src/main/java/com/bitmovin/analytics/license/FeatureConfigContainer.kt
@@ -1,0 +1,5 @@
+package com.bitmovin.analytics.license
+
+import com.bitmovin.analytics.features.errordetails.ErrorDetailTrackingConfig
+
+data class FeatureConfigContainer(val errorSegments: ErrorDetailTrackingConfig?)